LUCIDDB-QA: fix for LER-5721 - add counter to get allocated page count
[git-p4: depot-paths = "//open/lu/dev/": change = 9983]

+++ fennel/common/FileStatsTarget.cpp
@@ -54,7 +54,7 @@ void FileStatsTarget::endSnapshot()
     snapshotStream.close();
 }
 
-void FileStatsTarget::writeCounter(std::string name,uint value)
+void FileStatsTarget::writeCounter(std::string name,int64_t value)
 {
     snapshotStream << name << ' ' << value << std::endl;
 }
